Update Dependency from "diskutils" to "packaging" for Python 3.12 Fix file descriptor leak in service on quit(). Fixed a typo in __init__.py Update the download URL for the chromedriver package Prevent an error on already closed process.

+++ undetected_chromedriver/patcher.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # this module is part of undetected_chromedriver
 
-from distutils.version import LooseVersion
+from packaging import version
 import io
 import json
 import logging
@@ -94,8 +94,10 @@ class Patcher(object):
         # Set the correct repository to download the Chromedriver from
         if self.is_old_chromedriver:
             self.url_repo = "https://chromedriver.storage.googleapis.com"
+            self.download_url = self.url_repo + "/%s/%s"
         else:
             self.url_repo = "https://googlechromelabs.github.io/chrome-for-testing"
+            self.download_url = "https://storage.googleapis.com/chrome-for-testing-public/%s/%s/%s"
 
         self.version_main = version_main
         self.version_full = None
@@ -133,12 +135,13 @@ class Patcher(object):
         if self.user_multi_procs:
             with Lock():
                 files = list(p.rglob("*chromedriver*"))
-                most_recent = max(files, key=lambda f: f.stat().st_mtime)
-                files.remove(most_recent)
-                list(map(lambda f: f.unlink(), files))
-                if self.is_binary_patched(most_recent):
-                    self.executable_path = str(most_recent)
-                    return True
+                if len(files) > 0:
+                    most_recent = max(files, key=lambda f: f.stat().st_mtime)
+                    files.remove(most_recent)
+                    list(map(lambda f: f.unlink(), files))
+                    if self.is_binary_patched(most_recent):
+                        self.executable_path = str(most_recent)
+                        return True
 
         if executable_path:
             self.executable_path = executable_path
@@ -173,7 +176,7 @@ class Patcher(object):
             pass
 
         release = self.fetch_release_number()
-        self.version_main = release.version[0]
+        self.version_main = release.major
         self.version_full = release
         self.unzip_package(self.fetch_package())
         return self.patch()
@@ -233,14 +236,14 @@ class Patcher(object):
         """
         Gets the latest major version available, or the latest major version of self.target_version if set explicitly.
         :return: version string
-        :rtype: LooseVersion
+        :rtype: packaging.version.Version
         """
         # Endpoint for old versions of Chromedriver (114 and below)
         if self.is_old_chromedriver:
             path = f"/latest_release_{self.version_main}"
             path = path.upper()
             logger.debug("getting release number from %s" % path)
-            return LooseVersion(urlopen(self.url_repo + path).read().decode())
+            return version.parse(urlopen(self.url_repo + path).read().decode())
 
         # Endpoint for new versions of Chromedriver (115+)
         if not self.version_main:
@@ -251,7 +254,7 @@ class Patcher(object):
                 response = conn.read().decode()
 
             last_versions = json.loads(response)
-            return LooseVersion(last_versions["channels"]["Stable"]["version"])
+            return version.parse(last_versions["channels"]["Stable"]["version"])
 
         # Fetch the latest minor version of the major version provided
         path = "/latest-versions-per-milestone-with-downloads.json"
@@ -260,14 +263,14 @@ class Patcher(object):
             response = conn.read().decode()
 
         major_versions = json.loads(response)
-        return LooseVersion(major_versions["milestones"][str(self.version_main)]["version"])
+        return version.parse(major_versions["milestones"][str(self.version_main)]["version"])
 
     def parse_exe_version(self):
         with io.open(self.executable_path, "rb") as f:
             for line in iter(lambda: f.readline(), b""):
                 match = re.search(rb"platform_handle\x00content\x00([0-9.]*)", line)
                 if match:
-                    return LooseVersion(match[1].decode())
+                    return version.parse(match[1].decode())
 
     def fetch_package(self):
         """
@@ -277,11 +280,12 @@ class Patcher(object):
         """
         zip_name = f"chromedriver_{self.platform_name}.zip"
         if self.is_old_chromedriver:
-            download_url = "%s/%s/%s" % (self.url_repo, self.version_full.vstring, zip_name)
+            url_parts = (self.version_full.base_version, zip_name)
         else:
             zip_name = zip_name.replace("_", "-", 1)
-            download_url = "https://storage.googleapis.com/chrome-for-testing-public/%s/%s/%s"
-            download_url %= (self.version_full.vstring, self.platform_name, zip_name)
+            url_parts = (self.version_full.base_version, self.platform_name, zip_name)
+
+        download_url = self.download_url % url_parts
 
         logger.debug("downloading from %s" % download_url)
         return urlretrieve(download_url)[0]
